+++ package.json
@@ -1,38 +1,56 @@
 {
-  "name": "frourio-app",
+  "name": "frourio-demo",
   "version": "0.1.0",
-  "private": true,
+  "description": "Demo application of Frourio.",
+  "keywords": [
+    "frourio"
+  ],
+  "homepage": "https://github.com/technote-space/frourio-demo",
+  "bugs": {
+    "url": "https://github.com/technote-space/frourio-demo/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/technote-space/frourio-demo.git"
+  },
+  "license": "MIT",
+  "author": {
+    "name": "Technote",
+    "email": "technote.space@gmail.com",
+    "url": "https://technote.space"
+  },
   "scripts": {
-    "setup": "yarn setup:install && yarn setup:db && yarn build",
-    "setup:install": "yarn install && cd server && yarn install",
-    "setup:db": "yarn --cwd server setup && yarn migrate:seed",
-    "dev": "npm run migrate:dev && run-p dev:*",
-    "dev:client": "next dev -p 3000",
-    "dev:server": "npm run dev --prefix server",
-    "dev:aspida": "aspida --watch",
-    "build": "run-p build:client build:server",
+    "analyze": "ANALYZE=true next build",
+    "build": "run-s build:server build:client",
     "build:client": "aspida && prisma generate --schema server/prisma/schema.prisma && next build && next export",
     "build:server": "npm run build --prefix server",
     "build:types": "aspida && npm run build:frourio --prefix server",
-    "analyze": "ANALYZE=true next build",
+    "cover": "jest --coverage --verbose -i",
+    "cover:front": "jest --coverage --verbose -i src",
+    "cover:server": "jest --coverage --verbose server",
+    "cover:update": "jest --coverage --verbose -u -i src",
+    "dev": "npm run migrate:dev && run-p dev:*",
+    "dev:aspida": "aspida --watch",
+    "dev:client": "next dev -p 3000",
+    "dev:server": "npm run dev --prefix server",
     "lint": "eslint --ext .ts,.js,.tsx .",
     "lint:fix": "npm run lint -- --fix",
-    "migrate:init": "npm run migrate:init --prefix server",
     "migrate:dev": "npm run migrate:dev --prefix server",
+    "migrate:init": "npm run migrate:init --prefix server",
     "migrate:reset": "npm run migrate:reset --prefix server",
     "migrate:seed": "yarn ts-node ./server/prisma/seeds/seeder.ts",
+    "setup": "yarn install && yarn setup:db && yarn build",
+    "setup:db": "yarn --cwd server setup && yarn migrate:seed",
+    "setup:install": "yarn install && cd server && yarn install",
     "start": "run-p start:*",
     "start:client": "next start",
     "start:server": "npm start --prefix server",
-    "cover": "jest --coverage --verbose -i",
-    "cover:server": "jest --coverage --verbose server",
-    "cover:front": "jest --coverage --verbose -i src",
-    "cover:update": "jest --coverage --verbose -u -i src",
     "test": "npm run lint && npm run build:types && npm run cover",
     "typecheck": "npm run build:types && tsc --noEmit && tsc --noEmit -p server",
-    "update": "yarn update:front && yarn update:server",
+    "update": "run-s update:*",
     "update:front": "npx npm-check-updates -u --timeout 100000 -x @date-io/date-fns,next-transpile-modules && yarn install && yarn upgrade && yarn audit",
-    "update:server": "cd server/prisma && npx npm-check-updates -u --timeout 100000 && yarn install && yarn upgrade && yarn audit"
+    "update:server": "cd server/prisma && npx npm-check-updates -u --timeout 100000 && yarn install && yarn upgrade && yarn audit",
+    "update:electron": "cd src/electron && npx npm-check-updates -u --timeout 100000 && yarn install && yarn upgrade && yarn audit"
   },
   "dependencies": {
     "@aspida/axios": "^1.2.1",
