+++ server/package.json
@@ -3,18 +3,18 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "setup": "npm run setup:config && npm run setup:db",
-    "setup:config": "npx cpy-cli --no-overwrite --rename=.env .env.example ./ && npx cpy-cli --no-overwrite --rename=.env prisma/.env.example prisma/",
-    "setup:db": "npx del-cli prisma/dev.db && yarn migrate:dev",
-    "dev": "npm run migrate:dev && run-p dev:*",
-    "dev:server": "webpack --watch --mode=development",
-    "dev:frourio": "frourio --watch",
-    "dev:prisma": "prisma generate --watch",
     "build": "npm run build:frourio && webpack --mode=production",
     "build:frourio": "npm run setup:config && npm run migrate:dev && prisma generate && frourio",
-    "migrate:init": "prisma migrate dev --name init --preview-feature",
+    "dev": "npm run migrate:dev && run-p dev:*",
+    "dev:frourio": "frourio --watch",
+    "dev:prisma": "prisma generate --watch",
+    "dev:server": "webpack --watch --mode=development",
     "migrate:dev": "prisma migrate dev --preview-feature",
+    "migrate:init": "prisma migrate dev --name init --preview-feature",
     "migrate:reset": "prisma migrate reset --preview-feature",
+    "setup": "yarn install && run-s setup:*",
+    "setup:config": "npx cpy-cli --no-overwrite --rename=.env .env.example ./ && npx cpy-cli --no-overwrite --rename=.env prisma/.env.example prisma/",
+    "setup:db": "npx del-cli prisma/dev.db && yarn migrate:dev",
     "start": "cross-env NODE_ENV=production node index.js"
   },
   "dependencies": {
